Explicitly check the push host URL before attempting to push a gem
`gem` will immediately stop with no error if it encounters an unexpected
response when pushing a gem. To avoid this, I'm adding an explicit pre-flight
check that replicates the request `gem` will make, but wuth no authentication,
and testing for an 'Authentication required' response.

If the gem host URL has the wrong scheme or excess slashes, the gem host
usually responds with a redirect.
https://github.com/rubygems/rubygems/issues/4458

fixes https://github.com/fac/dev-platform/issues/202